Adding docstyle checking to pre-commit
The conventions ignored are:
D100	Missing docstring in public module
D104  Missing docstring in public package
D200	One-line docstring should fit on one line with quotes

Reasoning for D100 and D104 is that often these doc strings are beyond self-evident from the naming of the class or module. For D200 this is not always the case especially if we are stating the maxline length is 88.